backend/requirements.txt
# SkillSetu Backend - Python FastAPI + SQLite
# SQLite needs NO extra driver — it's built into Python!

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database - SQLite (built into Python, zero setup!)
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI - Google Gemini (Free)
google-generativeai==0.7.2
groq==0.9.0

# PDF generation (Resume)
reportlab==4.2.2

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# File handling
aiofiles==23.2.1
Pillow==11.0.0

# Notifications (optional)
twilio==9.2.2

# Utilities
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.1
email-validator==2.2.0
